updated more component names to reflect Flow
