+++ firefly/src/components/MyAccountPage.js
@@ -3,7 +3,7 @@ import Container from "@material-ui/core/Container";
 import { makeStyles } from "@material-ui/core/styles";
 import Paper from "@material-ui/core/Paper";
 import Grid from "@material-ui/core/Grid";
-import "../styles/accountPage.scss";
+import "../styles/MyAccountPage.scss";
 import { Link } from "react-router-dom";
 
 import image1 from "../assets/icons/Firefly.svg";
