+++ firefly/src/views/DevMenu.js
@@ -19,10 +19,10 @@ import ColorSliderTest from "../components/ColorSlider/ColorSliderTest";
 import SignInPage from "../components/SignInPage";
 import StepperTest from "../components/Steppers/StepperTest";
 import MultiStepRegistration from "../components/MultiStepRegistration";
-import AccountPage from "../components/AccountPage";
+import MyAccountPage from "../components/MyAccountPage";
 import Accessories from "../components/childProfile/Accessories";
-import EditProfile from "../components/CreateAndEditProfiles/EditProfile";
-import ActuallyCreateProfile from "../components/CreateAndEditProfiles/ActuallyCreateProfile";
+import EditProfilePage from "../components/CreateAndEditProfiles/EditProfilePage";
+import AddANewProfilePage from "../components/CreateAndEditProfiles/AddANewProfilePage";
 
 export default function Layout() {
   return (
@@ -54,15 +54,11 @@ export default function Layout() {
           />
           <Route exact path="/slider-test" component={ColorSliderTest} />
           <Route exact path="/stepper-test" component={StepperTest} />
-          <Route exact path="/account" component={AccountPage} />
+          <Route exact path="/account" component={MyAccountPage} />
           <Route exact path="/accessories" component={Accessories} />
 
-          <Route exact path="/edit-profile" component={EditProfile} />
-          <Route
-            exact
-            path="/create-profile"
-            component={ActuallyCreateProfile}
-          />
+          <Route exact path="/edit-profile" component={EditProfilePage} />
+          <Route exact path="/create-profile" component={AddANewProfilePage} />
         </Switch>
       </main>
     </div>
